Add basic framework for HTTP server.
The goal here is to make it really easy to create a standalone
HTTP server while still allowing for some flexibility in what
the HTTP server is used for.  This HTTP server is pretty easy
to instantiate and run but lays a framework that should also
allow developers building from it to easily use Servlet or
JaxRS handlers to customize it to their needs.